move AssertionError retry, flake8

+++ conda_build/environ.py
@@ -569,19 +569,23 @@ def get_install_actions(prefix, index, specs, config, retries=0):
                 raise DependencyNeedsBuildingError(exc)
             except (SystemExit, PaddingError, LinkError, DependencyNeedsBuildingError,
                     CondaError, AssertionError) as exc:
-                if 'lock' in str(exc) or isinstance(exc, AssertionError):
+                if 'lock' in str(exc):
                     log.warn("failed to get install actions, retrying.  exception was: %s",
                              str(exc))
                 elif ('requires a minimum conda version' in str(exc) or
-                        'link a source that does not' in str(exc)):
+                        'link a source that does not' in str(exc) or
+                        isinstance(exc, AssertionError)):
                     locks = utils.get_conda_operation_locks(config)
                     with utils.try_acquire_locks(locks, timeout=config.timeout):
                         pkg_dir = str(exc)
                         while os.path.dirname(pkg_dir) not in pkgs_dirs:
+                            input_dir = pkg_dir
                             pkg_dir = os.path.dirname(pkg_dir)
+                            if pkg_dir == input_dir:
+                                break
                         log.warn("I think conda ended up with a partial extraction for %s.  "
                                     "Removing the folder and retrying", pkg_dir)
-                        if os.path.isdir(pkg_dir):
+                        if pkg_dir in pkgs_dirs and os.path.isdir(pkg_dir):
                             utils.rm_rf(pkg_dir)
                 if retries < config.max_env_retry:
                     log.warn("failed to get install actions, retrying.  exception was: %s",
